add useCase

+++ dws-presentation/src/main/java/es/cesguiro/controller/admin/BookAdminController.java
@@ -40,7 +40,7 @@ public class BookAdminController {
 
         int total = bookAdminService.count();
 
-        PaginatedResponse<BookCollection> response = new PaginatedResponse<>(books, (int) total, page, pageSize, baseUrl + URL);
+        PaginatedResponse<BookCollection> response = new PaginatedResponse<>(books, total, page, pageSize, baseUrl + URL);
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
